Initial cut
AeronKeyValueClient has a lot of obviously bugs in it around corelating load requests. There's clearly no guarantee that stores occur either (no plans to address this)

The demo client cycles through keys A and B which do not initial exist and update them to be one of three values. It cycles these values over time. No assertions are given on the new value.